Move buttons and adapt imports
