+++ client/src/pages/BrowsePage.js
@@ -5,7 +5,7 @@ import { useQuery } from "react-query";
 import ImagePreview from "../components/ImagePreview";
 import { Link } from "react-router-dom";
 import FavoriteIcon from "@material-ui/icons/Favorite";
-import { FaveButton } from "../components/Button";
+import { FaveButton } from "../components/Buttons";
 import TitlePreview from "../components/TitlePreview";
 import Header from "../components/Header";
 import Navbar from "../components/Navbar";
